tianzhongbo: symlink support

+++ mfschunkserver/hddspacemgr.cc
@@ -3364,6 +3364,8 @@ static int hdd_int_duptrunc(uint64_t chunkid,uint32_t version,uint32_t newversio
 
 static int hdd_int_delete(uint64_t chunkid,uint32_t version) {
 	chunk *c;
+	int blen = 256;
+	char buf[256];
 	c = hdd_chunk_find(chunkid);
 	if (c==NULL) {
 		return ERROR_NOCHUNK;
@@ -3371,6 +3373,10 @@ static int hdd_int_delete(uint64_t chunkid,uint32_t version) {
 	if (c->version!=version && version>0) {
 		hdd_chunk_release(c);
 		return ERROR_WRONGVERSION;
+	}
+	if ((blen = readlink(c->filename, buf, blen)) != -1) {
+		buf[blen] = 0;
+		unlink(buf);
 	}
 	if (unlink(c->filename)<0) {
 		hdd_error_occured(c);	// uses and preserves errno !!!
@@ -3623,7 +3629,7 @@ static inline void hdd_add_chunk(folder *f,const char *fullname,uint64_t chunkid
 	if (c == NULL) return;
 	if (c->filename!=NULL) {	// already have this chunk
 		if (version <= c->version) {	// current chunk is older
-			if (todel<2) { // this is R/W fs?
+			if (version < c->version && todel<2) { // this is R/W fs?
 				unlink(fullname); // if yes then remove file
 			}
 		} else {
